% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{auth}
\alias{auth}
\title{auth}
\usage{
auth()
}
\description{
Interactive authenticator interface. Similar in scope to Google
Authenticator. To use, you must first set up your database of keys
using \code{otpdb()}.
}
\details{
Only the requested database key will be decrypted in memory. Keys are never
decrypted on disk.
}
\references{
\url{https://en.wikipedia.org/wiki/Google_Authenticator}
}
\seealso{
\code{\link{otpdb}}
}
