% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotp.r
\name{hotp}
\alias{hotp}
\title{hotp}
\usage{
hotp(key, counter, digits = 6)
}
\arguments{
\item{key}{The secret key.}

\item{counter}{A "counter" (some integer).}

\item{digits}{The number of digits of the return.}
}
\value{
An integer with \code{digits} digits.
}
\description{
Implementation of the HMAC-based One-Time Password algorithm.
}
\examples{
rotp::totp("asdf", 10)

}
\references{
\url{https://en.wikipedia.org/wiki/HMAC-based_One-time_Password_algorithm}
}
