% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otpdb.r
\name{otpdb}
\alias{otpdb}
\title{otpdb}
\usage{
otpdb()
}
\description{
Interactive interface for the secret key database. After setting up your
keys using this interface, you can authenticate with \code{auth()}.
}
\details{
The database is located at \code{~/.rotpdb}. On Windows, this will be on your
"My Documents" folder.

Secret keys are encrypted using envelope encryption. You will need an RSA key
which the openssl package can find. You can check for your keypair via
\code{openssl::my_pubkey()}.
}
\seealso{
\code{\link{auth}}
}
