% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotp-package.R
\docType{package}
\name{rotp-package}
\alias{rotp-package}
\title{rotp}
\description{
Time-based One-Time Password (TOTP) is a relatively common 2-factor
scheme, with Google 'Authenticator' being a well-known implementation. This
provides the same technique for R together with an encrypted database
schema.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
