% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totp.r
\name{totp}
\alias{totp}
\title{totp}
\usage{
totp(key, interval = 30, digits = 6)
}
\arguments{
\item{key}{The secret key.}

\item{interval}{The interval of time in seconds.}

\item{digits}{The number of digits of the return.}
}
\value{
An integer with \code{digits} digits.
}
\description{
Implementation of the Time-based One-Time Password algorithm.
}
\examples{
rotp::totp("asdf")

}
\references{
\url{https://en.wikipedia.org/wiki/Time-based_One-time_Password_algorithm}
}
